Moved tuple-return test to scrypto-test project
