+++ radix-engine-tests/assets/blueprints/Cargo.toml
@@ -57,7 +57,6 @@ members = [
     "transaction_limits",
     "transaction_runtime",
     "tx_processor_access",
-    "tuple-return",
     "events",
     "events_invalid",
     "logger",
