+++ scrypto-test/tests/blueprints/Cargo.toml
@@ -1,0 +1,12 @@
+[workspace]
+members = [
+    "tuple-return"
+]
+
+[profile.release]
+opt-level = 'z'        # Optimize for size.
+lto = true             # Enable Link Time Optimization.
+codegen-units = 1      # Reduce number of codegen units to increase optimizations.
+panic = 'abort'        # Abort on panic.
+strip = true           # Strip the symbols.
+overflow-checks = true # Panic in the case of an overflow.
