+++ scrypto-test/tests/blueprints/tuple-return/src/lib.rs
